Split TestUpcall into subclasses
