+++ test/jdk/ProblemList.txt
@@ -533,7 +533,7 @@ java/beans/XMLEncoder/Test6570354.java 8015593 macosx-all
 # jdk_foreign
 
 java/foreign/TestMismatch.java 8249684 macosx-all
-java/foreign/TestUpcall.java#stack 8275584 macosx-aarch64
+java/foreign/TestUpcallBase.java#stack 8275584 macosx-aarch64
 
 ############################################################################
